refactor: move src/routes to src/app/routes
