+++ apps/analog-app/src/app/routes/product-details.component.html
