+++ apps/analog-app/src/app/routes/products.[productId].ts
@@ -2,8 +2,8 @@ import { Component, OnInit } from '@angular/core';
 import { ActivatedRoute } from '@angular/router';
 import { CommonModule } from '@angular/common';
 
-import { Product, products } from '../app/products';
-import { CartService } from '../app/cart.service';
+import { Product, products } from '../products';
+import { CartService } from '../cart.service';
 
 @Component({
   selector: 'app-product-details',
