+++ apps/analog-app/src/app/routes/shipping/index.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { Component, OnInit } from '@angular/core';
 
 import { Observable } from 'rxjs';
-import { CartService } from '../../app/cart.service';
+import { CartService } from '../../cart.service';
 
 @Component({
   selector: 'app-shipping',
