+++ apps/analog-app/tsconfig.app.json
@@ -5,6 +5,6 @@
     "types": []
   },
   "files": ["src/main.ts", "src/polyfills.ts"],
-  "include": ["src/**/*.d.ts", "src/routes/**/*.ts"],
+  "include": ["src/**/*.d.ts", "src/app/routes/**/*.ts"],
   "exclude": ["**/*.test.ts", "**/*.spec.ts"]
 }
